newmarker and location shifting maindata.py
Experimental futures which is tested and working:

plots all players' cities and adds one more marker to simulate unit production. Then shifts it through a flask script to simulate movement. Current problem; it adds a js code to move the specific marker. Since it is not practical to inject a new code for every produced unit/marker , js code should get the js variable name from flask and locate it in the page code and change it's coordinates.